Add ASIOne, OpenRouter and OpenAI plugins
