+++ providers/lib_llm_ext.py
@@ -144,156 +144,6 @@ class AIProvider(AbstractAIProvider):
         return text.replace("_quote_", '"').replace("_apostrophe_", "'").replace("</arg_value>", " ") \
                     .replace("</tool_call>", " ").replace("<arg_value>", " ").replace("<tool_call>", " ")
 
-class OpenRouterProvider(AIProvider):
-    """OpenRouter provider with reasoning mode enabled (reasoning tokens excluded from the response)."""
-
-    def _create_client(self) -> Optional[openai.OpenAI]:
-        """Create OpenRouter client from environment."""
-        proxy_url = os.environ.get("GATEWAY_URL")
-        if proxy_url:
-            base_url = f"{proxy_url.rstrip('/')}/openrouter/"
-            logger.info(f"[lib_llm_ext.OpenRouterProvider._create_client]: Connecting via proxy: {base_url}")
-            return openai.OpenAI(
-                    api_key="proxy",
-                    base_url=base_url,
-                    )
-        if self._var_name in os.environ:
-            return openai.OpenAI(api_key=os.environ.get(self._var_name), base_url=self._base_url)
-
-        return None
-    
-    def _openrouter_extra_body(self, content: str, max_tokens: int) -> Dict[str, Any]:
-        sysmsg, _ = _split_system_user(content)
-
-        body = {
-            "reasoning": {
-                "enabled": True,
-                "max_tokens": max_tokens,
-                "exclude": True,
-            }
-        }
-
-        # Helps OpenRouter sticky-route requests for better cache locality.
-        # Keep this stable per agent/session.
-        session_id = os.environ.get("OPENROUTER_SESSION_ID")
-        if not session_id and sysmsg:
-            session_id = _stable_cache_key("openrouter", self._model_name, sysmsg)
-
-        if session_id:
-            body["session_id"] = session_id[:256]
-
-        model = self._model_name.lower()
-
-        # OpenRouter supports top-level cache_control for Anthropic Claude routes.
-        if model.startswith("anthropic/"):
-            body["cache_control"] = {
-                "type": "ephemeral",
-                "ttl": os.environ.get("OPENROUTER_CACHE_TTL", "5m"),
-            }
-
-        return body
-
-
-    def chat(self, content: str, max_tokens: int = 6000, reasoning: str = "medium", **kwargs) -> str:
-        extra_body = _merge_dicts(
-            self._openrouter_extra_body(content, max_tokens),
-            kwargs.pop("extra_body", None),
-        )
-
-        return super().chat(
-            content=content,
-            max_tokens=max_tokens,
-            reasoning=reasoning,
-            extra_body=extra_body,
-            **kwargs,
-        )
-
-class AsiOneProvider(AIProvider):
-    """Lazy AI provider with on-demand initialization."""
-
-    def __init__(self, name: str, var_name: str, model_name: str, base_url: str):
-        super().__init__(name, var_name, model_name, base_url)
-
-    def chat(self, content: str, max_tokens: int = 6000, reasoning: str = "medium", **kwargs) -> str:
-        """Send chat request, initializing client if needed."""
-        self._ensure_client()
-
-        if self._client is None:
-            raise RuntimeError(f"{self.name} not configured (set {self._var_name})")
-
-        sysmsg, usermsg = content.split(":-:-:-:")
-        try:
-            response = self._client.chat.completions.create(
-                model=self._model_name,
-                messages=[{"role": "system", "content": sysmsg},
-                          {"role": "user", "content": usermsg}],
-                max_tokens=max_tokens,
-                extra_body={
-                    "enable_thinking": True,
-                    "thinking_budget": 6000 
-                },
-                **kwargs
-            )
-
-            raw = response.choices[0].message.content
-            _log_raw(self._name, self._model_name, raw)
-            resp = self._clean_text(raw)
-            return resp
-        except Exception as e:
-            logger.exception(f"[lib_llm_ext.ASIOneProvider.chat]: Exception while communicating with LLM: {e}")
-            return ""
-
-
-class OpenAIProvider(AIProvider):
-    """OpenAI provider using the Responses API (reasoning models)."""
-
-    def chat(self, content: str, max_tokens: int = 6000, reasoning: str = "medium", **kwargs) -> str:
-        """Send chat request via the Responses API, initializing client if needed."""
-        self._ensure_client()
-
-        if self._client is None:
-            raise RuntimeError(f"{self.name} not configured (set {self._var_name})")
-
-        sysmsg, usermsg = _split_system_user(content)
-
-        try:
-            create_kwargs = {
-                "instructions": sysmsg,
-                "model": self._model_name,
-                "input": usermsg,
-                "max_output_tokens": max_tokens,
-                "reasoning": {"effort": reasoning},
-                "prompt_cache_key": os.environ.get("OPENAI_PROMPT_CACHE_KEY", _stable_cache_key("openai", self._model_name, sysmsg)),
-            }
-            # GPT-5.5 supports only 24h; GPT-5.4 also supports extended retention.
-            if self._model_name.startswith(("gpt-5.5", "gpt-5.4")):
-                create_kwargs["prompt_cache_retention"] = "24h"
-
-            create_kwargs.update(kwargs)
-
-            response = self._client.responses.create(**create_kwargs)
-
-            usage = getattr(response, "usage", None)
-            if usage:
-                input_tokens = getattr(usage, "input_tokens", None)
-                output_tokens = getattr(usage, "output_tokens", None)
-                total_tokens = getattr(usage, "total_tokens", None)
-                details = getattr(usage, "input_tokens_details", None)
-                cached_tokens = getattr(details, "cached_tokens", None) if details else None
-
-                print(
-                    f"[LLM_USAGE] provider={self._name} model={self._model_name} "
-                    f"input_tokens={input_tokens} output_tokens={output_tokens} "
-                    f"total_tokens={total_tokens} cached_tokens={cached_tokens}"
-                )
-
-            raw = response.output_text or ""
-            _log_raw(self._name, self._model_name, raw)
-            return self._clean_text(raw)
-        except Exception as e:
-            logger.exception(f"[lib_llm_ext.OpenAIProvider.chat]: Exception while communicating with LLM: {e}")
-            return ""
-
 
 class TestProvider(AbstractAIProvider):
     """Test provider for mocking LLM output"""
@@ -334,14 +184,7 @@ def _get_provider(name: str) -> Optional[AIProvider]:
 
 
 # Register all providers (cheap - just stores config)
-_register_provider(name="ASICloud", var_name="ASI_API_KEY", model_name="minimax/minimax-m3", base_url="https://inference.asicloud.cudos.org/v1")
-_register_provider(name="Anthropic", var_name="ANTHROPIC_API_KEY", model_name="claude-opus-4-8", base_url="https://api.anthropic.com/v1/")
-_register_provider(name="Ollama-local", var_name="OLLAMA_API_KEY", model_name="qwen3.5:9b", base_url="http://localhost:11434/v1")
-_register_provider_instance(AsiOneProvider(name="ASIOne", var_name="ASIONE_API_KEY", model_name="asi1-ultra", base_url="https://api.asi1.ai/v1"))
-_register_provider_instance(OpenRouterProvider(name="OpenRouter", var_name="OPENROUTER_API_KEY", model_name="z-ai/glm-5.2", base_url="https://openrouter.ai/api/v1"))
-_register_provider_instance(OpenRouterProvider(name="MiniMaxM3", var_name="OPENROUTER_API_KEY", model_name="minimax/minimax-m3", base_url="https://openrouter.ai/api/v1"))
 _register_provider_instance(TestProvider())
-_register_provider_instance(OpenAIProvider(name="OpenAI", var_name="OPENAI_API_KEY", model_name="gpt-5.5", base_url="https://api.openai.com/v1"))
 
 
 def callProvider(provider_name: str, content: str, max_tokens: int = 6000, reasoning: str = "medium") -> str:
